Bug fix and organising some code
